Module Emile.List

val of_string_with_crlf : string -> (t list, [> error ]) result

of_string_with_crlf s parses s which can be a list of named group or a single mailbox separated by a comma. In the case of a group, s starts with a name and contains a list of email separated by a comma and terminates with a semicolon:

Gallium: Gabriel <gabriel@gallium.fr>, Armael <armael@gallium.fr>;

s must terminate with CRLF as the delimiter. If the parser fails, it return an error error.

val of_string : string -> (t list, [> error ]) result

of_string s is of_string_with_crlf but did not need CRLF at the end. It's possible that of_string did not consume all s.

val of_string_raw : off:int -> len:int -> ?tmp:Bigstringaf.t -> string -> (int * t list, [> error ]) result

of_string_raw s off len is of_string_with_crlf but did not need CRLF. It parses only a sub-part of s starting at off and computes at most len bytes. It returns how many bytes it consumed.

If the user has an already allocated Bigstringaf.t, it can use it as an internal buffer to parse given input s.

val to_string : t list -> string

to_string lst returns a well-formed string which represents the given list of t.